move duplicated logic into validator
